+++ Engine/Engine.vcxproj
@@ -66,33 +66,34 @@
     <ClCompile Include="world\Texture.cpp" />
   </ItemGroup>
   <ItemGroup>
-    <ClInclude Include="cameras\Camera.h" />
-    <ClInclude Include="cameras\Controller.h" />
-    <ClInclude Include="Engine.h" />
-    <ClInclude Include="errors.h" />
-    <ClInclude Include="graphics\Config.h" />
-    <ClInclude Include="graphics\Geometry.h" />
-    <ClInclude Include="graphics\Graphics.h" />
-    <ClInclude Include="graphics\Math.h" />
-    <ClInclude Include="graphics\Shaders.h" />
-    <ClInclude Include="graphics\Vertex.h" />
-    <ClInclude Include="Logger.h" />
-    <ClInclude Include="renderer\Renderer.h" />
-    <ClInclude Include="renderer\VulkanBuffer.h" />
-    <ClInclude Include="renderer\VulkanCommandBuffer.h" />
-    <ClInclude Include="renderer\VulkanCommandPool.h" />
-    <ClInclude Include="renderer\VulkanDescriptor.h" />
-    <ClInclude Include="renderer\VulkanDevice.h" />
-    <ClInclude Include="renderer\VulkanInstance.h" />
-    <ClInclude Include="renderer\VulkanPipeline.h" />
-    <ClInclude Include="renderer\VulkanRenderPass.h" />
-    <ClInclude Include="renderer\VulkanSwapchain.h" />
-    <ClInclude Include="Window.h" />
-    <ClInclude Include="world\GameObject.h" />
-    <ClInclude Include="world\Material.h" />
-    <ClInclude Include="world\Mesh.h" />
-    <ClInclude Include="world\Scene.h" />
-    <ClInclude Include="world\Texture.h" />
+    <ClInclude Include="Includes\camera\Camera.h" />
+    <ClInclude Include="Includes\camera\Controller.h" />
+    <ClInclude Include="Includes\Engine.h" />
+    <ClInclude Include="Includes\errors.h" />
+    <ClInclude Include="Includes\graphics\Config.h" />
+    <ClInclude Include="Includes\graphics\Geometry.h" />
+    <ClInclude Include="Includes\graphics\Graphics.h" />
+    <ClInclude Include="Includes\graphics\Math.h" />
+    <ClInclude Include="Includes\graphics\Shaders.h" />
+    <ClInclude Include="Includes\graphics\Vertex.h" />
+    <ClInclude Include="Includes\Logger.h" />
+    <ClInclude Include="Includes\renderer\Renderer.h" />
+    <ClInclude Include="Includes\renderer\VulkanBuffer.h" />
+    <ClInclude Include="Includes\renderer\VulkanCommandBuffer.h" />
+    <ClInclude Include="Includes\renderer\VulkanCommandPool.h" />
+    <ClInclude Include="Includes\renderer\VulkanDescriptor.h" />
+    <ClInclude Include="Includes\renderer\VulkanDevice.h" />
+    <ClInclude Include="Includes\renderer\VulkanInstance.h" />
+    <ClInclude Include="Includes\renderer\VulkanPipeline.h" />
+    <ClInclude Include="Includes\renderer\VulkanRenderPass.h" />
+    <ClInclude Include="Includes\renderer\VulkanSwapchain.h" />
+    <ClInclude Include="Includes\Window.h" />
+    <ClInclude Include="Includes\world\GameObject.h" />
+    <ClInclude Include="Includes\world\Geometry.h" />
+    <ClInclude Include="Includes\world\Material.h" />
+    <ClInclude Include="Includes\world\Mesh.h" />
+    <ClInclude Include="Includes\world\Scene.h" />
+    <ClInclude Include="Includes\world\Texture.h" />
   </ItemGroup>
   <ItemGroup>
     <None Include="assets\shaders\HLSL\no_texture_shader.frag" />
