+++ Engine/Engine.vcxproj.filters
@@ -98,86 +98,89 @@
     </ClCompile>
   </ItemGroup>
   <ItemGroup>
-    <ClInclude Include="Logger.h">
-      <Filter>Header</Filter>
+    <ClInclude Include="Includes\world\GameObject.h">
+      <Filter>Engine\World</Filter>
     </ClInclude>
-    <ClInclude Include="renderer\Renderer.h">
+    <ClInclude Include="Includes\world\Geometry.h">
+      <Filter>Engine\World</Filter>
+    </ClInclude>
+    <ClInclude Include="Includes\world\Material.h">
+      <Filter>Engine\World</Filter>
+    </ClInclude>
+    <ClInclude Include="Includes\world\Mesh.h">
+      <Filter>Engine\World</Filter>
+    </ClInclude>
+    <ClInclude Include="Includes\world\Scene.h">
+      <Filter>Engine\World</Filter>
+    </ClInclude>
+    <ClInclude Include="Includes\world\Texture.h">
+      <Filter>Engine\World</Filter>
+    </ClInclude>
+    <ClInclude Include="Includes\renderer\Renderer.h">
       <Filter>Engine\Renderer\Header</Filter>
     </ClInclude>
-    <ClInclude Include="renderer\VulkanBuffer.h">
+    <ClInclude Include="Includes\renderer\VulkanBuffer.h">
       <Filter>Engine\Renderer\Header</Filter>
     </ClInclude>
-    <ClInclude Include="renderer\VulkanCommandBuffer.h">
+    <ClInclude Include="Includes\renderer\VulkanCommandBuffer.h">
       <Filter>Engine\Renderer\Header</Filter>
     </ClInclude>
-    <ClInclude Include="renderer\VulkanCommandPool.h">
+    <ClInclude Include="Includes\renderer\VulkanCommandPool.h">
       <Filter>Engine\Renderer\Header</Filter>
     </ClInclude>
-    <ClInclude Include="renderer\VulkanDescriptor.h">
+    <ClInclude Include="Includes\renderer\VulkanDescriptor.h">
       <Filter>Engine\Renderer\Header</Filter>
     </ClInclude>
-    <ClInclude Include="renderer\VulkanDevice.h">
+    <ClInclude Include="Includes\renderer\VulkanDevice.h">
       <Filter>Engine\Renderer\Header</Filter>
     </ClInclude>
-    <ClInclude Include="renderer\VulkanInstance.h">
+    <ClInclude Include="Includes\renderer\VulkanInstance.h">
       <Filter>Engine\Renderer\Header</Filter>
     </ClInclude>
-    <ClInclude Include="renderer\VulkanPipeline.h">
+    <ClInclude Include="Includes\renderer\VulkanPipeline.h">
       <Filter>Engine\Renderer\Header</Filter>
     </ClInclude>
-    <ClInclude Include="renderer\VulkanRenderPass.h">
+    <ClInclude Include="Includes\renderer\VulkanRenderPass.h">
       <Filter>Engine\Renderer\Header</Filter>
     </ClInclude>
-    <ClInclude Include="renderer\VulkanSwapchain.h">
+    <ClInclude Include="Includes\renderer\VulkanSwapchain.h">
       <Filter>Engine\Renderer\Header</Filter>
     </ClInclude>
-    <ClInclude Include="errors.h">
-      <Filter>Header</Filter>
+    <ClInclude Include="Includes\graphics\Config.h">
+      <Filter>Engine\Graphics</Filter>
     </ClInclude>
-    <ClInclude Include="cameras\Camera.h">
+    <ClInclude Include="Includes\graphics\Geometry.h">
+      <Filter>Engine\Graphics</Filter>
+    </ClInclude>
+    <ClInclude Include="Includes\graphics\Graphics.h">
+      <Filter>Engine\Graphics</Filter>
+    </ClInclude>
+    <ClInclude Include="Includes\graphics\Math.h">
+      <Filter>Engine\Graphics</Filter>
+    </ClInclude>
+    <ClInclude Include="Includes\graphics\Shaders.h">
+      <Filter>Engine\Graphics</Filter>
+    </ClInclude>
+    <ClInclude Include="Includes\graphics\Vertex.h">
+      <Filter>Engine\Graphics</Filter>
+    </ClInclude>
+    <ClInclude Include="Includes\camera\Camera.h">
       <Filter>Engine\Cameras</Filter>
     </ClInclude>
-    <ClInclude Include="world\GameObject.h">
-      <Filter>Engine\World</Filter>
-    </ClInclude>
-    <ClInclude Include="world\Material.h">
-      <Filter>Engine\World</Filter>
-    </ClInclude>
-    <ClInclude Include="world\Mesh.h">
-      <Filter>Engine\World</Filter>
-    </ClInclude>
-    <ClInclude Include="world\Scene.h">
-      <Filter>Engine\World</Filter>
-    </ClInclude>
-    <ClInclude Include="world\Texture.h">
-      <Filter>Engine\World</Filter>
-    </ClInclude>
-    <ClInclude Include="cameras\Controller.h">
+    <ClInclude Include="Includes\camera\Controller.h">
       <Filter>Engine\Cameras</Filter>
     </ClInclude>
-    <ClInclude Include="graphics\Vertex.h">
-      <Filter>Engine\Graphics</Filter>
-    </ClInclude>
-    <ClInclude Include="graphics\Graphics.h">
-      <Filter>Engine\Graphics</Filter>
-    </ClInclude>
-    <ClInclude Include="graphics\Config.h">
-      <Filter>Engine\Graphics</Filter>
-    </ClInclude>
-    <ClInclude Include="graphics\Math.h">
-      <Filter>Engine\Graphics</Filter>
-    </ClInclude>
-    <ClInclude Include="Window.h">
+    <ClInclude Include="Includes\Engine.h">
       <Filter>Header</Filter>
     </ClInclude>
-    <ClInclude Include="Engine.h">
+    <ClInclude Include="Includes\errors.h">
       <Filter>Header</Filter>
     </ClInclude>
-    <ClInclude Include="graphics\Shaders.h">
-      <Filter>Engine\Graphics</Filter>
+    <ClInclude Include="Includes\Logger.h">
+      <Filter>Header</Filter>
     </ClInclude>
-    <ClInclude Include="graphics\Geometry.h">
-      <Filter>Engine\Graphics</Filter>
+    <ClInclude Include="Includes\Window.h">
+      <Filter>Header</Filter>
     </ClInclude>
   </ItemGroup>
   <ItemGroup>
