+++ ExampleCube/ExampleCube.vcxproj
@@ -116,7 +116,7 @@
       <SDLCheck>true</SDLCheck>
       <PreprocessorDefinitions>_DEBUG;_CONSOLE;%(PreprocessorDefinitions)</PreprocessorDefinitions>
       <ConformanceMode>true</ConformanceMode>
-      <AdditionalIncludeDirectories>$(VULKAN_SDK)\Include;$(SolutionDir)\dependencies\glm;$(SolutionDir)\dependencies\glfw-3.3.4\include;$(SolutionDir)\dependencies\stb_image;$(SolutionDir)\dependencies\assimp-5.0.1\include;$(SolutionDir)\Engine;%(AdditionalIncludeDirectories)</AdditionalIncludeDirectories>
+      <AdditionalIncludeDirectories>$(VULKAN_SDK)\Include;$(SolutionDir)\dependencies\glm;$(SolutionDir)\dependencies\glfw-3.3.4\include;$(SolutionDir)\dependencies\stb_image;$(SolutionDir)\dependencies\assimp-5.0.1\include;$(SolutionDir)\Engine\Includes;%(AdditionalIncludeDirectories)</AdditionalIncludeDirectories>
     </ClCompile>
     <Link>
       <SubSystem>Console</SubSystem>
